here

+++ data/migrations/20231230235937_create-products.js
@@ -1,32 +1,48 @@
 exports.up = function(knex) {
-  return knex.schema
-    .createTable("category", (table) => {
-      table.increments().primary();
-      table.string("category_name", 255).notNullable();
-    })
-    .createTable("subcategory", (table) => {
-      table.increments();
-      table.string("subcategory_name", 255).notNullable();
-      table
-        .integer("categoryId")
-        .notNullable()
-        .references("id")
-        .inTable("category")
-        .onDelete("CASCADE"); // Add this line to delete subcategories when a category is deleted
-    })
-    .createTable("products", (table) => {
-      table.increments();
-      table.string("product_name", 255).notNullable();
-      table.string("product_description", 1000).notNullable();
-      table.string("product_price", 255).notNullable();
-      table
-        .integer("subcategoryId")
-        .nullable()
-        .references("id")
-        .inTable("subcategory")
-        .onDelete("CASCADE"); // Add this line to delete products when a subcategory is deleted
+  return knex.schema.hasTable('category').then((exists) => {
+    if (!exists) {
+      return knex.schema.createTable("category", (table) => {
+        table.increments().primary();
+        table.string("category_name", 255).notNullable();
+      });
+    }
+  })
+  .then(() => {
+    return knex.schema.hasTable('subcategory').then((exists) => {
+      if (!exists) {
+        return knex.schema.createTable("subcategory", (table) => {
+          table.increments();
+          table.string("subcategory_name", 255).notNullable();
+          table
+            .integer("categoryId")
+            .notNullable()
+            .references("id")
+            .inTable("category")
+            .onDelete("CASCADE");
+        });
+      }
     });
+  })
+  .then(() => {
+    return knex.schema.hasTable('products').then((exists) => {
+      if (!exists) {
+        return knex.schema.createTable("products", (table) => {
+          table.increments();
+          table.string("product_name", 255).notNullable();
+          table.string("product_description", 1000).notNullable();
+          table.string("product_price", 255).notNullable();
+          table
+            .integer("subcategoryId")
+            .nullable()
+            .references("id")
+            .inTable("subcategory")
+            .onDelete("CASCADE");
+        });
+      }
+    });
+  });
 };
+
 
 /**
  * @param { import("knex").Knex } knex
